Refactored solution structure
The solution now had 3 additional projects: Game, Headless and
CopyResources. Game creates a project for the client. Headless is to
create a server and CopyResources is a workaround to force copy the
resources in the res folder to the output dir. The project deletes
itself after being build so that the scripts get executed every time.

I've also added property sheets since I am statically linking. I did not
think about dynamic linking and don't feel like refactoring all the
functions.

Finally, I've added two scripts, one to delete a folder and one to copy
a folder to another folder.