Moved VoxelCraft into proj folder and optimized
Fixed the build for debug. There was a weird issue with moving
initializer lists in debug build. Changing it to a vector solved the
problem.

Optimized render pipeline a bit. When generating a mesh for a chunk
the chunk checks if it is enclosed by 6 others. If it is, it checks
if they are only air. If they are, the chunk doesn't generate since
it isn't visible.

The mesh also only submits itself to the renderer if it isn't empty
which saves on OpenGL calls.